[Qwen4Exp] Fix QSA warmup for fused output gate, make gate mandatory
warmup_qsa_sparse_paged_attention still used the pre-fusion kernel
signatures, so every .warmup() call misaligned positionally and raised
TypeError at startup with enable_jit_warmup=True (the default). Pass the
gate tensor and its strides in both warmup calls.

attn_output_gate is hardcoded on for this architecture, so make
output_gate a required keyword argument end to end, drop the None
specialization branches from both kernels, and assert gate contiguity.
This leaves exactly one kernel specialization per profile, which is the
one warmup compiles.

Match the PLE kernel's final residual add to eager fp32 opmath so the
fused result is bitwise identical to the unfused path; tighten the
fused-conv test from atol/rtol=3e-2 to exact equality.

Tested on GB200:
- tests/models/qwen4_exp/test_ple.py: 36 passed
- tests/models/qwen4_exp/test_qsa_reference.py: 59 passed
- warmup_qsa_sparse_paged_attention smoke: all 7 profiles compile

+++ vllm/models/qwen4_exp/nvidia/ops/ple.py
@@ -452,7 +452,7 @@ def _ple_conv_kernel(
         mask=c_mask,
         other=0.0,
     )
-    ple_output = outer_residual + ple_output
+    ple_output = outer_residual.to(tl.float32) + ple_output.to(tl.float32)
     if launch_pdl:
         tl.extra.cuda.gdc_launch_dependents()
     tl.store(
